requirements.txt
python-dotenv
langchain-openai
langchain-chroma
langchain-experimental
docling
flask
flask-restful
flask-cors